+++ data-extraction/src/js/Extractors/index.ts
@@ -1,21 +1,5 @@
-import { getDataExtractorApi } from "../DataExtractorApi";
-import { TypeScriptAstDataExtractor } from "./TypeScriptDataExtractors";
-import { AsIsDataExtractor } from "./AsIsDataExtractor";
-import { GetDebugVisualizationDataExtractor } from "./GetDebugVisualizationDataExtractor";
+export { TypeScriptAstDataExtractor } from "./TypeScriptDataExtractors";
+export { AsIsDataExtractor } from "./AsIsDataExtractor";
+export { GetDebugVisualizationDataExtractor } from "./GetDebugVisualizationDataExtractor";
 
-export {
-	AsIsDataExtractor,
-	TypeScriptAstDataExtractor,
-	GetDebugVisualizationDataExtractor,
-};
-
-export function registerAll() {
-	const api = getDataExtractorApi();
-	for (const item of [
-		new TypeScriptAstDataExtractor(),
-		new AsIsDataExtractor(),
-		new GetDebugVisualizationDataExtractor(),
-	]) {
-		api.registerExtractor(item);
-	}
-}
+export { registerDefaultDataExtractors } from "./registerDefaultDataExtractors";