+++ data-extraction/src/js/createGraph.ts
@@ -4,47 +4,10 @@ import {
 	NodeGraphData,
 } from "../CommonDataTypes";
 
-export function createGraphFromPointers<T>(
-	roots: Record<string, T | undefined | null>,
-	infoSelector: (
-		item: T
-	) => {
-		id?: string | number;
-		edges: ({ to: T } & Omit<EdgeGraphData, "from" | "to">)[];
-	} & Omit<NodeGraphData, "id">
-): CommonDataTypes.Graph {
-	const marker = {};
-
-	interface Pointer {
-		marker: {};
-		name: string;
-		value: T | null | undefined;
-	}
-
-	const items = Object.entries(roots).map<Pointer>(([name, value]) => ({
-		marker,
-		name,
-		value,
-	}));
-
-	const includeLabelInName = false;
-
-	return createGraph<T | Pointer>(items, item => {
-		if ("marker" in item && item["marker"] === marker) {
-			return {
-				id: "label____" + item.name,
-				color: "orange",
-				label: item.name,
-				edges: [{ to: item.value!, color: "orange", label: "" }].filter(
-					t => !!t.to
-				),
-			};
-		} else {
-			return infoSelector(item as T);
-		}
-	});
-}
-
+/**
+ * Given a list of roots, it creates a graph by following their edges recursively.
+ * Tracks cycles.
+ */
 export function createGraph<T>(
 	roots: T[],
 	infoSelector: (
@@ -102,4 +65,47 @@ export function createGraph<T>(
 		}
 	}
 	return r;
+}
+
+/**
+ * Given a labeled list of roots, it creates a graph by following their edges recursively.
+ * Tracks cycles.
+ */
+export function createGraphFromPointers<T>(
+	roots: Record<string, T | undefined | null>,
+	infoSelector: (
+		item: T
+	) => {
+		id?: string | number;
+		edges: ({ to: T } & Omit<EdgeGraphData, "from" | "to">)[];
+	} & Omit<NodeGraphData, "id">
+): CommonDataTypes.Graph {
+	const marker = {};
+
+	interface Pointer {
+		marker: {};
+		name: string;
+		value: T | null | undefined;
+	}
+
+	const items = Object.entries(roots).map<Pointer>(([name, value]) => ({
+		marker,
+		name,
+		value,
+	}));
+
+	return createGraph<T | Pointer>(items, item => {
+		if ("marker" in item && item["marker"] === marker) {
+			return {
+				id: "label____" + item.name,
+				color: "orange",
+				label: item.name,
+				edges: [{ to: item.value!, color: "orange", label: "" }].filter(
+					t => !!t.to
+				),
+			};
+		} else {
+			return infoSelector(item as T);
+		}
+	});
 }