+++ example-project/src/demo_custom-data-extractor.ts
@@ -1,0 +1,34 @@
+import {
+	registerDefaultDataExtractors,
+	getDataExtractorApi,
+} from "@hediet/debug-visualizer-data-extraction";
+
+// Registers all existing extractors.
+registerDefaultDataExtractors();
+
+getDataExtractorApi().registerExtractor({
+	id: "my-extractor",
+	getExtractions: (data, collector) => {
+		if (data instanceof Foo) {
+			collector.addExtraction({
+				id: "my-extractor",
+				name: "My Extractor",
+				priority: 2000,
+				extractData: () => ({ kind: { text: true }, text: "Foo" }),
+			});
+		}
+	},
+});
+
+setTimeout(() => {
+	new Main().run();
+}, 0);
+
+class Foo {}
+
+class Main {
+	run() {
+		const f = new Foo();
+		debugger;
+	}
+}