+++ extension/CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 0.9.0
 
 -   Initial release
+
+## Current
+
+-   Minor refactoring of `@hediet/debug-visualizer-data-extraction`'s API
+-   Improved documentation.